SSO Modifications and Documentation (#103)
This is an implementation and documentation of configuring python social auth with Nautobot. It includes the documentation on how to configure SSO for various sources, and some required changes to fix the LOGIN_URL, the navigation header to use the correct URL if SSO is enabled, add the optional dependencies, and adds the URL patterns if SSO is enabled.