Update signals for cables to always call rebuild_paths, even on cable creation.
This handles the case where creating a path through a circuit doesn't update
the first cable, when a second cable is connected.  This mimics the same
action that is done when creating path through front/rear ports.  This allows
a circuit to be both a Path Endpoint for cases such as DSL or Fiber Broadband,
but also updates the cables when using the Circuit as a passthrough such as
dark fiber.