Refactor install steps to update .bashrc AFTER virtualenv creation
We moved some things around, as the virtualenv was previously created by
root, so that when one became `nautobot`, the `/opt/nautobot/bin`
directory was already there making the `$PATH` be set upon sudo.

This revises the instructions to instruct the reader to add the setting
of `NAUTOBOT_ROOT` in `.bashrc` and reloading with `source .bashrc`
until after the user creates the virtualenv as the `nautobot` user to
alleviate the confusion of the `$PATH` not being correct.