Fix circular import introduced by #322. (#344)
This is cannot reproduced with unit tests. The unit tests did not catch this issue. This adds a `InterfaceConnectionsListView .get_queryset()` method to emulate the built-in Django generic view API that is unfortunately not implemented in the Nautobot core that we inherited from NetBox. So this also adds an `__init__` method to the view to call `self.get_queryset()` on init. Short of refactoring all of the base generic views, this is the cleanest fix I could think of for now. 

I did attempt to convert `InterfaceConnectionsListView.queryset` to a property, and while that did alleviate the issue initially, it resulted in many tests failing because [`nautobot.utilities.views.ObjectPermissionRequiredMixin.has_permission` wants to replace `self.queryset` with `self.queryset.restrict(user, action)` on line 77](https://github.com/nautobot/nautobot/blob/develop/nautobot/utilities/views.py#L77).