Production Docker Image (#195)
This PR is an initial cut at a production docker container.  Some components of this PR:

- UWSGI is hosting the HTTP/HTTPS service in the container
- Configuration from environment vars a default configuration is used inside the container
- Docs
- Docker image caching is used to save time for an "integration tests" step which currently only starts the container and makes sure we are able to hit the UI

I added a couple of docker-compose overrides.  One is dev which I made the default, this is required for the volumes in the dev environment but not the build environment.  I created a build override as well which specifically runs the production container with no local volumes which could potentially be used for testing.  You can specify which override file to use with:

```
export OVERRIDE_FILENAME=docker-compose.build.yml
```

The docker-compose.dev.yml is the default.  This will work with the existing commands `invoke build`, `invoke start`, `invoke debug`, etc....

This PR does require the following env vars be set in the Travis CI environment `DOCKER_HUB_PASSWORD`, `DOCKER_HUB_USERNAME`, `GHCR_ACCESS_TOKEN` and `GHCR_USERNAME`

Co-authored-by: Jathan McCollum <jathan@gmail.com>
Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>