Fix default Redis settings to not have `None` as a password
The default settings from the config generated from template using `nautobot-server init` were changed in v1.0.1 to look for certain settings from environment variables. This was done to facilitate a 12-factor application pattern in coordination with publishing an official Docker image.

This fix corrects this to satisfy a working default configuration, and to support customizing settings using `NAUTOBOT_*` environment variables.