Fix issue with missing tags on MySQL (#550)
Had to change the base class for `TaggedItem` from `GenericTaggedItemBase` to `GenericUUIDTaggedItemBase`, to get the django-taggit machinery to do a UUID-based comparison of the prefetched "tags". Without that the prefetched values weren't matching correctly, and were returning empty. (See: https://github.com/jazzband/django-taggit/pull/663)

This does not require any migrations because `TaggedItem.object_id` was already changed to a `UUIDField` in a previous commit.

This also fixed an issue that was causing tags from emitting in the API entirely. We do not yet have integration tests for that.