Add custom validator to allow slug field to be blank.
This allows slug to not be marked as a required field,
even though it follows the unique_together constraint
on the model.  If a slug is not provided, one will be
generated by the AutoSlugField, if one is provided,
then the DRF will do the validations for uniqueness
and slug regex.