Show help only if static file exists.
This change uses the finders.find function from static
to detect if a static file exists for the help page that is
to be displayed. This will work whether using the default
storage or a seperate storage provider for static files.

This also includes integration tests to verify the link is
shown or hidden. I had to add views for AnotherExampleModel
in order to get this test to work.