Remove extra closing of tag
The small tag has a chance of being closed twice but is opened only once. This causes an error when trying to test HTML on a page that uses the created_updated template. It also causes the page that has both created and updated to show "Created on ..." to be small, but "Updated on..." to be normal size on the same line.