update non-core docs to use nautobot.apps.jobs module (#4767)
* update non-core docs to use nautobot.apps.jobs module

Updated the module that the jobs and app developer guides use in their
examples to nautobot.apps.jobs which is the modern location since
Nautobot 1.5.2. ref #4759 / #4460

* switch example from BaseJob to Job

Turns out in this case the usage of BaseJob is not correct and not
explained in the documentation so it's better to just be consistent and
use Job until the docs can contain a specific example of when BaseJob
should be used.

* update the Job testing for nautobot.apps module

Switched the Job testing docs and examples to using nautobot.apps where
possible while rewording the section to take a newer focus since users
can and should reference the 1.x version of the docs and not the 2.x
version for details about 1.x Nautobot.