fix: Fixed LLDP neighbor view where interface names have spaces
HTML IDs cannot, by standard, have spaces in their names. The LLDP neighbor view tab sets the output table row's ID to the interface name. If a device interface name has a space, this breaks the jquery lookup to find the interface row. This fix changes the lookup to use a `data-` attribute for the interface name and lookup.