feat: Added `get_http_session` to ExternalIntegration
This feature provides a method to create a `requests.Session` initialized with the relevant configuration provided by an `ExternalIntegration`. This includes potential auth tokens, basic auth, ssl verification and custom CA file path. The reasoning behind this feature is to provide a re-usable interface for app developers to get http sessions for API clients defined as ExternalIntegrations. For instance, to interface with AWX a client needs the URL, ssl verification info, token or username/password. These can all be defined in `ExternalIntegration` and `get_http_session` can then be used to initialize an API client.