Wireless UI (#6338)
* Initial UI commit

* Add the JSONArrayChoiceFormField

Also add templates and continue to flush out the UI.

* Fix JSONArrayChoiceFormField

This change allows the widget to select the proper choice on load, so that existing choices are selected.

* Update factory.py

channel_width and allowed_channel_list should contain unique values.

* Update forms.py

Remove unused imports.

* Add inline formset for assignments.

* Add missing fields for the forms

This adds the M2M fields into the create/edit form

* Fix verbose name for SSID

* Finish Wireless Views

This adds the formsets to the create/edit views as well as adding the tables to the table view.

* Fix issue with formsets not being able to use query_params

* Add Wireless tab to Device Detail View

* More UI Changes and Tests

* Fix Access Point Group on Device Edit

* Fix issue with tests

The field when errors are present in form uses `data=[]`, which is not a string and will error out on JSONFields `json.loads(data)`. This ensures that we always have a json valid string object.

* Fix test cases

Also adds bulk_edit for channel_width

* Fix ordering in NumericArrayField

* Address feedback from Sprint Review

* Ruff and change fragments.

* Update nautobot/wireless/templates/wireless/wirelessnetwork_retrieve.html

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>

* Update nautobot/wireless/templates/wireless/supporteddatarate_retrieve.html

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>

* Update nautobot/wireless/templates/wireless/wirelessnetwork_retrieve.html

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>

* Address some of the Peer Review feedback

* Fix issue with field errors not rendering red.

* Add Form Validation for the Formset.

* Uncommit accidentally committed code

* Address BulkEdit Form issues.

* Address Additional feedback

* Update tables.py

Ruff

* Fix filter test

* Remove unused Code

* SupportedDataRate Tests should not conflict with factory.

* Fix bug in Formset lookup

We only care about the closest TR that has a class applied (which all dynamic formsets would). If the class attribute doesn't exist, it isn't a dynamic formset.

* Fix channel_width on RadioProfileTestCases

* Fix Controller Active tabs for wireless

* Address feedback on Paginator

* Apply suggestions from code review

Co-authored-by: Gary Snider <75227981+gsnider2195@users.noreply.github.com>

* Fix factory generation of devices for Wireless Access Points

* Add additional filters requested

* Ruff

* Move JobQueueFactory to before Tenant Factory

This is to try to get around the Tenant filter tests, This should have been before the TenantFactory.

* Ensure 2 JobQueue's have tenants

---------

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>
Co-authored-by: Gary Snider <75227981+gsnider2195@users.noreply.github.com>