fixed prefix/circuit count for location detail view stats panel (#7812)
Without this change `SELECT DISTINCT` is not used for the count status
for the prefix and circuit models, causing stats to be inflated if we
are looking at a location that has at least two children with
overlapping related prefixes or circuits. For VLANs, there was already a
distinct present, but in order to optimize for performance I have added
a condition to only apply `distinct` to the three relevant querysets if
we are actually looking at more than one location.

I have applied parts of #7810 in order to make this possible (namely the
`list` for eager queryset evaluation).

Co-authored-by: Ken Celenza <ken@celenza.org>