This change allows modifying job forms through job_form_wrapper instead of block content
The use case here (and used within import_objects.html) is to modify the nav-tabs to be within a card-header.

Additionally, this moves the warning when has_sensitive_variables to the more appropriate job_execution card. This also removes the unnecessary `job` tab at the top of the job page.