This change allows modifying job forms through job_form_wrapper inste… (#7840)
* This change allows modifying job forms through job_form_wrapper instead of block content

The use case here (and used within import_objects.html) is to modify the nav-tabs to be within a card-header.

Additionally, this moves the warning when has_sensitive_variables to the more appropriate job_execution card. This also removes the unnecessary `job` tab at the top of the job page.

* Revert change on sensitive_variables message

Turns out that the whole job_execution card is hidden when has_sensitive_variables is set.

* Move sensitive variables to it's own card.

This way any inheritance and override of job_form_wrapper does not need to worry about the sensitive variable message.

* Remove card-body from sensitive-variables-message add header.

* Cleanup

* Update nautobot/docs/development/jobs/job-structure.md

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>

* Update nautobot/extras/templates/extras/job.html

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>

* Create 7839.fixed

* Apply suggestions

---------

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>